triagesync_backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
channels>=4.1.0
python-dotenv>=1.0.1
dj-database-url>=2.2.0
psycopg[binary]>=3.2.1
gunicorn>=21.2.0
whitenoise>=6.6.0
google-generativeai>=0.4.0
